+++ include/open62541pp/Logger.h
@@ -5,6 +5,10 @@
 
 namespace opcua {
 
+// forward declarations
+class Client;
+class Server;
+
 /**
  * Log level.
  * @see UA_LogLevel
@@ -34,6 +38,12 @@ enum class LogCategory {
 
 /// Log function signature.
 using Logger = std::function<void(LogLevel, LogCategory, std::string_view msg)>;
+
+/// Generate log message with client's logger.
+void log(Client& client, LogLevel level, LogCategory category, std::string_view msg);
+
+/// Generate log message with servers's logger.
+void log(Server& server, LogLevel level, LogCategory category, std::string_view msg);
 
 /* -------------------------------------- Utility functions ------------------------------------- */